# packages/shared/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://127.0.0.1:8080
  description: Generated server url
paths:
...
  /api/v1/withdrawals:
    post:
      tags:
      - earnings-controller
      operationId: withdraw
      parameters:
      - name: Idempotency-Key
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalReq"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/assignment/{id}/confirm-join:
    post:
      tags:
      - assignment-controller
      operationId: confirmJoin
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/assignment/recommend:
    post:
      tags:
      - assignment-controller
      operationId: recommend
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/assignment/confirm:
    post:
      tags:
      - assignment-controller
      operationId: confirm
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/assignment/cancel:
    post:
      tags:
      - assignment-controller
      operationId: cancel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CancelRequest"
        required: true
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/order-references:
    get:
      tags:
      - earnings-controller
      operationId: orders
      parameters:
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: memberNo
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListMapStringObject"
...
  /api/v1/earnings/summary:
    get:
      tags:
      - earnings-controller
      operationId: earnings
      parameters:
      - name: memberNo
        in: query
        required: true
        schema:
          type: string
      - name: projectId
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
...
  /api/v1/assignment:
    get:
      tags:
      - assignment-controller
      operationId: list_10
      parameters:
      - name: projectId
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
      - name: groupId
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListMapStringObject"
  /api/v1/assignment/{id}:
    get:
      tags:
      - assignment-controller
      operationId: get_4
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseMapStringObject"
  /api/v1/assignment/pending:
    get:
      tags:
      - assignment-controller
      operationId: pending
      parameters:
      - name: projectId
        in: query
        required: false
        schema:
          type: string
      responses:
        "500":
          description: Internal Server Error
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseVoid"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ApiResponseListMapStringObject"
...
components:
  schemas:
    ApiResponseVoid:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
...
    ApiResponseMapStringObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: object
          additionalProperties:
            type: object
    WithdrawalReq:
      required:
      - amount
      - memberNo
      - method
      type: object
      properties:
        memberNo:
          type: string
        amount:
          type: number
        method:
          type: string
        accountInfo:
          type: string
...
    RecommendRequest:
      required:
      - memberNo
      - projectId
      type: object
      properties:
        memberNo:
          type: string
        projectId:
          type: string
    ConfirmRequest:
      required:
      - assignmentId
      type: object
      properties:
        assignmentId:
          type: integer
          format: int64
        groupId:
          type: string
        overrideReason:
          type: string
    CancelRequest:
      required:
      - assignmentId
      type: object
      properties:
        assignmentId:
          type: integer
          format: int64
        reason:
          type: string
...
    ApiResponseListMapStringObject:
      type: object
      properties:
        code:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: array
          items:
            type: object
            additionalProperties:
              type: object
...